@@ -1,4 +1,4 @@
-import { Switch, Route } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Header from "./Header";
 import Hero from "./Hero";
 import Home from "./Home";
@@ -8,28 +8,20 @@ import PowerEditForm from "./PowerEditForm";
 
 function App() {
   return (
-    <div>
-      <Header />
-      <main>
-        <Switch>
-          <Route exact path="/hero_powers/new">
-            <HeroPowerForm />
-          </Route>
-          <Route exact path="/powers/:id/edit">
-            <PowerEditForm />
-          </Route>
-          <Route exact path="/powers/:id">
-            <Power />
-          </Route>
-          <Route exact path="/heroes/:id">
-            <Hero />
-          </Route>
-          <Route exact path="/">
-            <Home />
-          </Route>
-        </Switch>
-      </main>
-    </div>
+    <Router>
+      <div>
+        <Header />
+        <main>
+          <Routes>
+            <Route path="/hero_powers/new" element={<HeroPowerForm />} />
+            <Route path="/powers/:id/edit" element={<PowerEditForm />} />
+            <Route path="/powers/:id" element={<Power />} />
+            <Route path="/heroes/:id" element={<Hero />} />
+            <Route path="/" element={<Home />} />
+          </Routes>
+        </main>
+      </div>
+    </Router>
   );
 }
 
